moved UserCategoryFollow model to users app
