+++ theinkspot/users/tests/test_models.py
@@ -1,6 +1,7 @@
 import pytest
+from django.db import IntegrityError
 
-from theinkspot.users.models import User
+from theinkspot.users.models import User, UserCategoryFollow
 
 pytestmark = pytest.mark.django_db
 
@@ -74,3 +75,47 @@ class TestUserModel:
     def test_create_super_user(self, superuser):
         is_user_a_superuser = superuser.is_superuser
         assert is_user_a_superuser is True
+
+
+@pytest.mark.django_db
+class TestUserCategoryFollow:
+    def test_UserCategoryFollow_unique_together(self):
+        unique_together = UserCategoryFollow._meta.unique_together
+        assert unique_together[0] == ("user", "category")
+
+    def test_user_follow_category(self, user, category):
+        follow = UserCategoryFollow.objects.create(user=user, category=category)
+        assert user.followers.count() == 1
+        assert category.followed_categories.count() == 1
+        assert follow.get_email is False
+
+    def test_get_email_label(self):
+        label = UserCategoryFollow._meta.get_field("get_email").verbose_name
+        assert label == "get notified about this category"
+
+    def test_get_email_default_value(self):
+        default = UserCategoryFollow._meta.get_field("get_email").default
+        assert default is False
+
+    def test_user_unfollow_category(self, user, category):
+        UserCategoryFollow.objects.create(user=user, category=category)
+        user.followers.first().delete()
+        assert user.followers.count() == 0
+        assert category.followed_categories.count() == 0
+
+    def test_user_follow_already_followed_category(self, user, category):
+        UserCategoryFollow.objects.create(user=user, category=category)
+
+        with pytest.raises(IntegrityError):
+            UserCategoryFollow.objects.create(user=user, category=category)
+
+    def test_user_unfollow_already_unfollowed_category(self, user, category):
+        UserCategoryFollow.objects.create(user=user, category=category)
+        user.followers.first().delete()
+
+        with pytest.raises(AttributeError):
+            user.followers.first().delete()
+
+    def test_object_name_is_user_follows_category(self, user, category):
+        follow = UserCategoryFollow.objects.create(user=user, category=category)
+        assert str(follow) == f"{follow.user} follows {follow.category}"
